pyproject.toml
[tool.poetry]
name = "llm-learning-from-experience"
version = "0.1.0"
description = ""
authors = ["Hanan Shteingart <hshteingart@nvidia.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
pandas = "^2.2.3"
scipy = "^1.15.1"
scikit-learn = "^1.6.1"
seaborn = "^0.13.2"
matplotlib = "^3.10.0"
plotly = "^6.0.0"
langchain = "^0.3.17"
openai = "^1.61.0"
pytest = "^8.3.4"
langchain-community = "^0.3.16"
langchain-openai = "^0.3.3"
python-dotenv = "^1.0.1"
langgraph = "^0.2.69"
toml = "^0.10.2"
tqdm = "^4.67.1"
loguru = "^0.7.3"
scalene = "^1.5.51"
joblib = "^1.4.2"
retry = "^0.9.2"
fire = "^0.7.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# add src pythonpath to pytest
[tool.pytest.ini_options]
pythonpath = ["src"]
